CBD-4603: Tweak generation to allow building from a different version
Introduced DockerfileVariant.TargetVersion, which is the name of the
directory under enterprise/PRODUCT. This will generally be the same
as Version, but occasionally we must build the Docker image with
version X from a released product with a slightly different version.

Also simplified generateVersions() to make it easier to add in
specific exception cases.