MB-52093: Build multi-arch images for Server 7.1.1+
Allow for Server Dockerfiles to be multi-arch. In particular, this means
we need to bake in multiple SHA256 checksums and switch between them at
"docker build" time based on the architecture being built, and similarly
determine the installer filename to download at build time.

For SHA256, I added a bit of shell scripting to choose between two new
Dockerfile build-args, CB_SHA256_amd64 and CB_SHA256_arm64, which are
only declared when creating a multi-arch Dockerfile. For older
single-arch Dockerfiles, there is just a single CB_SHA256 build-arg as
before. I also added a new CB_SKIP_CHECKSUM build-arg which allows
skipping the sha256sum check entirely; this will make it easier to
handle non-production Docker builds such as toys.

For the installer filename, there is still only a single CB_PACKAGE
build-arg, same as previously. However when doing a multi-arch build,
the Dockerfile will replace @@ARCH@@ in that filename with the
build-time architecture. This again was to keep things easier for
non-production use cases, which can always simply specify the CB_PACKAGE
build-arg regardless of whether it's a multi-arch Dockerfile or not.

Remove the previous workaround that involved creating multiple
Dockerfiles for multiple architectures, in particular because aarch64
was only supported on amzn2 (an rpm-based distro). That workaround was
only necessary for Server 7.1.0, and made the generator program flow
harder to follow, so I've removed it entirely. This made it possible to
get rid of the multi-stage amazonlinux:2 Dockerfile which had to build
runit from source.

I also re-created the 7.1.0 Dockerfiles and scripts using the new
generator. This is slightly questionable since those no longer match
what the current images were made from, but the new generator will keep
wanting to create the new multi-arch form for 7.1.0 which will get
annoying. I made a new branch on github.com/couchbase/docker called
"server-7.1.0-generator" from the master commit prior to my changes,
which we can use if we ever need to re-do the 7.1.0 Dockerfiles.

I could have also eliminated PKG_COMMAND and the related args to switch
between "apt-get" and "yum", but I left those in place since they don't
hurt readability much and they'll be handy if we someday want to roll
the RHCC Dockerfile generation into this repository.