reconfigure default ubuntu version to 22.0.4 for couchbase-server and couchbase-columnar.
Failure of python-httplib2 mentioned in https://github.com/couchbase/docker/pull/216 was due
to cb-docker-tool landing on an inaccurate couchbase-server version.  This was a special use
case for creating toy columnar container.  I have modified the jenkins job to use
couchbase-columnar template.  It is safe to use 22.0.4 as default now.

-Ming