MB-65496: Fix PROFILE handling for couchbase-columnar
Columnar builds >= 1.2.0 do not want `/etc/couchbase.d/config_profile`
in the image, but < 1.2.0 still requires that file to exist with the
contents `columnar`.

Switch the Dockerfile.template handling to match couchbase-server, where
PROFILE is a template argument that creates the block to create that
file only when the argument is set. Modify the generator to set
PROFILE=columnar for couchbase-columnar builds when the version is
< 1.2.0.